+++ CHANGELOG.md
@@ -5,5 +5,7 @@
 ## 4.0.0
 
 ### 新特性
+* 新增Object类，用于实现MongoDB的ID生成策略
+* 验证码单独成为一个模块hutool-captcha
 
 ### Bug修复